requirements.txt
requests
selenium
Pillow